pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jwt_token_x21124213"
version = "0.0.6"
authors = [
  { name="Rajendra Chavan", email="rajendrachavan.jvm@gmail.com" },
]
description = "Json Web Token generator and authenticator library"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'PyJWT ~= 2.6.0'
]

[project.urls]
"Homepage" = "https://github.com/rajendrachavan/jwt-token-python"
"Bug Tracker" = "https://github.com/rajendrachavan/jwt-token-python/issues"